+++ nextgisweb/webmap/test/test_api.py
@@ -1,13 +1,12 @@
 import pytest
 import transaction
 
-from nextgisweb.models import DBSession
 from nextgisweb.auth import User
-
+from nextgisweb.models import DBSession
 from nextgisweb.webmap.model import WebMap, WebMapItem
 
 ANNOTATION_SAMPLE = dict(
-    description='1', geom='POINT (0 0)',
+    description='1', geom='POINT (0 0)', public=True,
     style=dict(string='string', int=0, bool=True, none=None)
 )